refactor: PoolMap logic
- moved PoolMap logic into its own custom hook, reducing file size in PoolMap component
